% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALCPHENOTYPE.R
\name{predictionAccuracybyCV}
\alias{predictionAccuracybyCV}
\title{Calculate Cross-Validation Scores using OncoPredict}
\usage{
predictionAccuracybyCV(
  trainingExprData,
  trainingPtype,
  testExprData,
  batchCorrect,
  powerTransformPhenotype = TRUE,
  removeLowVaryingGenes = 0.2,
  minNumSamples,
  selection = 1,
  printOutput,
  pcr = FALSE,
  removeLowVaringGenesFrom,
  report_pc = FALSE,
  cc = FALSE,
  percent = 80,
  rsq = FALSE,
  folder = FALSE,
  cvFold = -1,
  parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{trainingExprData}{The training data. A matrix of expression levels. rownames() are genes, colnames() are samples (cell line names or cosmic ides, etc.). rownames() must be specified and must contain the same type of gene ids as "testExprData"}

\item{trainingPtype}{The known phenotype for "trainingExprData". This can be a one-drug vector with one value per training sample or a matrix with training samples as rows and drugs or phenotypes as columns. This matrix can contain NA values, that is ok (they are removed in the calcPhenotype() function).}

\item{testExprData}{The test data where the phenotype will be estimated. It is a matrix of expression levels, rows contain genes and columns contain samples, "rownames()" must be specified and must contain the same type of gene ids as "trainingExprData".}

\item{batchCorrect}{How should training and test data matrices be homogenized. Choices are "eb" (default) for ComBat, "qn" for quantile normalization, "standardize" for within-dataset z-score standardization, "rank", "rank_then_eb", or "none" for no homogenization.}

\item{powerTransformPhenotype}{Should the phenotype be power transformed before we fit the regression model? Default to TRUE, set to FALSE if the phenotype is already known to be highly normal.}

\item{removeLowVaryingGenes}{What proportion of low varying genes should be removed? 20 percent be default}

\item{minNumSamples}{How many training and test samples are required. Print an error if below this threshold}

\item{selection}{How should duplicate gene ids be handled. Default is -1 which asks the user. 1 to summarize by their or 2 to disguard all duplicates.}

\item{printOutput}{Set to FALSE to supress output.}

\item{pcr}{Indicates whether or not you'd like to use pcr for feature (gene) reduction. Options are 'TRUE' and 'FALSE'. If you indicate 'report_pc=TRUE' you need to also indicate 'pcr=TRUE'}

\item{removeLowVaringGenesFrom}{Determine method to remove low varying genes. Options are 'homogenizeData' and 'rawData'.}

\item{report_pc}{Indicates whether you want to output the training principal components. Options are 'TRUE' and 'FALSE'.}

\item{cc}{Indicate if you want correlation coefficients for biomarker discovery.}

\item{percent}{Indicate percent variability (of the training data) you'd like principal components to reflect if pcr=TRUE. Default is 80 for 80\%
These are the correlations between a given gene of interest across all samples vs. a given drug response across samples.
These correlations can be ranked to obtain a ranked correlation to determine highly correlated drug-gene associations.}

\item{rsq}{Indicate whether or not you want to output the R^2 values for the data you train on from true and predicted values.
These values represent the percentage in which the optimal model accounts for the variance in the training data.
Options are 'TRUE' and 'FALSE'.}

\item{folder}{Retained for compatibility with calcPhenotype arguments. Cross-validation results are returned as an object.}

\item{cvFold}{Indicate the number of k-folds wanted in the CV calculation. -1 indicates a leave-one-out cross validation}

\item{parallel}{If TRUE, run cross-validation folds in parallel. The default is FALSE.}

\item{cores}{The number of cores to use when parallel is TRUE. Parallel execution uses forked processes via parallel::mclapply, which is not available for multicore execution on Windows PCs; on Windows, predictionAccuracybyCV will warn and run serially.}
}
\value{
A list containing cross-validated predicted phenotype values and real phenotype values.
}
\description{
This function predicts a phenotype (drug sensitivity score) when provided with microarray or bulk RNAseq gene expression data of different platforms.
The imputations are performed using ridge regression, training on a gene expression matrix where phenotype is already known.
This function integrates training and testing datasets via a user-defined procedure, and power transforming the known phenotype.
}
\examples{
set.seed(1)
genes <- paste0("gene", 1:30)
trainingExprData <- matrix(rnorm(30 * 8), nrow=30,
                           dimnames=list(genes, paste0("train", 1:8)))
trainingPtype <- matrix(rnorm(8), ncol=1,
                        dimnames=list(colnames(trainingExprData), "drug1"))
cvResult <- predictionAccuracybyCV(trainingExprData, trainingPtype,
                                   testExprData=NULL,
                                   batchCorrect="none",
                                   powerTransformPhenotype=FALSE,
                                   removeLowVaryingGenes=0,
                                   minNumSamples=0,
                                   selection=1,
                                   printOutput=FALSE,
                                   pcr=FALSE,
                                   removeLowVaringGenesFrom="rawData",
                                   cvFold=2)
names(cvResult)
}
\keyword{and}
\keyword{drug}
\keyword{phenotype}
\keyword{predict}
\keyword{sensitivity}
