% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALCPHENOTYPE.R
\name{summarizeGenesByMean}
\alias{summarizeGenesByMean}
\title{Average over duplicate gene values}
\usage{
summarizeGenesByMean(exprMat)
}
\arguments{
\item{exprMat}{A gene expression matrix with genes as rownames() and samples as colnames().}
}
\value{
A gene expression matrix that does not contain duplicate genes.
}
\description{
This function takes a gene expression matrix and if duplicate genes are measured, summarizes them by their means.
}
\examples{
exprMat <- matrix(seq_len(12), nrow=3,
                  dimnames=list(paste0("gene", 1:3), paste0("sample", 1:4)))
summarizeGenesByMean(exprMat)
}
\keyword{Summarize}
\keyword{by}
\keyword{duplicate}
\keyword{genes}
\keyword{mean.}
\keyword{their}
