% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDWAS.R
\name{idwas}
\alias{idwas}
\title{This function will idwas every drug against every CNV or somatic mutation for your cancer type.}
\usage{
idwas(drug_prediction, data, n = 10, cnv)
}
\arguments{
\item{drug_prediction}{The drug prediction data. Must be a data frame. rownames() are samples, and colnames() are drugs. Make sure that the rownames() are of the same form as the sample names in your cnv or mutation data.
e.g. if the rownames() are TCGA barcodes of the form TCGA-##-####-###, make sure your cnv/mutation data also uses samples in the form TCGA-##-####-###}

\item{data}{The cnv or mutation data. Must be a data frame. If you wish to use cnv data, use the output from map_cnv().
If you wish to use mutation data, use the method for downloading mutation data outlined in the vignette; if not, ensure your data file
includes the following columns: 'Variant_Classification', 'Hugo_Symbol', 'Tumor_Sample_Barcode'.}

\item{n}{The number of samples you want CNVs or mutations to be amplified in. The default is 10.}

\item{cnv}{TRUE or FALSE. Indicate whether or not you would like to test cnv data. If TRUE, you will test cnv data. If FALSE, you will test mutation data.}
}
\description{
This function will test every drug against every CNV or somatic mutation for your cancer type.
}
\keyword{CNV}
\keyword{idwas}
\keyword{data}
\keyword{genes.}
\keyword{mutation}
\keyword{or}
\keyword{to}
