% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDWAS.R
\name{map_cnv}
\alias{map_cnv}
\title{This function maps cnv data to genes. The output of this function is a .RData file called map.RData; this file contains theCnvQuantVecList_mat (rows are genes, and columns are samples) and tumorSamps (indicates which samples are primary tumor samples, 01A).}
\usage{
map_cnv(Cnvs)
}
\arguments{
\item{Cnvs}{The cnv data. A table with the following colnames: Sample (named using the TCGA patient barcode), Chromosome, Start, End, Num_Probes, and Segment_Mean.}
}
\value{
A .RData file called, map.RData, which stores two objects: theCnvQuantVecList_mat (rows are genes, columns are samples), tumorSamps (indicates which samples are primary tumor/01A). This output will serve as the input for test().
}
\description{
This function maps cnv data to genes. The output of this function is a .RData file called map.RData; this file contains theCnvQuantVecList_mat (rows are genes, and columns are samples) and tumorSamps (indicates which samples are primary tumor samples, 01A).
}
\keyword{CNV}
\keyword{Map}
\keyword{data}
\keyword{genes}
\keyword{to}
