% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALCPHENOTYPE.R
\name{summarizeGenesByMean}
\alias{summarizeGenesByMean}
\title{This function takes a gene expression matrix and if duplicate genes are measured, summarizes them by their means.}
\usage{
summarizeGenesByMean(exprMat)
}
\arguments{
\item{exprMat}{A gene expression matrix with genes as rownames() and samples as colnames().}
}
\value{
A gene expression matrix that does not contain duplicate genes.
}
\examples{
\donttest{
try(summarizeGenesByMean(exprMat))
}
}
\description{
This function takes a gene expression matrix and if duplicate genes are measured, summarizes them by their means.
}
\keyword{Summarize}
\keyword{by}
\keyword{duplicate}
\keyword{genes}
\keyword{mean.}
\keyword{their}
